添加进度条，样式调整

+++ RESTFUL/src/main/java/com/prs/hub/file/service/FileService.java
@@ -42,7 +42,7 @@ public interface FileService {
      * @param userReqDTO 用户信息
      * @return
      */
-    Integer saveFileDetail(String filePath,String fileName,UserReqDTO userReqDTO);
+    Long saveFileDetail(String filePath,String fileName,UserReqDTO userReqDTO);
 
 
 }
